Feture |BS-9| add pundit

+++ app/controllers/items_controller.rb
@@ -1,14 +1,14 @@
 class ItemsController < ApplicationController
   def create
     service = Cart::ItemCreateService.new(current_user, item_params, params)
-    service.call(current_order)
+    authorize service.call(current_order)
     session[:current_order_id] = service.order.id unless current_order
     redirect_message(I18n.t('controllers.added_in_cart'), :notice)
   end
 
   def destroy
     service = Cart::ItemDeleteService.new(current_order, params)
-    service.call
+    authorize service.call
     session.delete(:current_order_id) if service.delete_order?
     redirect_message(I18n.t('controllers.deleted_from_cart'), :notice)
   end
